alert confirm

+++ src/pages/detalle/detalle.component.ts
@@ -1,7 +1,8 @@
-import { ListaDeseosService } from '../../app/services/lista-deseos.service';
-import { NavController, NavParams } from 'ionic-angular';
-import { Lista, ListaItem } from '../../app/classes';
+import { AlertController, NavController, NavParams } from 'ionic-angular';
 import { Component, OnInit } from '@angular/core';
+
+import { Lista, ListaItem } from '../../app/classes';
+import { ListaDeseosService } from '../../app/services/lista-deseos.service';
 
 @Component({
     selector: 'app-detalle',
@@ -11,9 +12,13 @@ export class DetalleComponent implements OnInit {
 
     lista: Lista[];
     index: number;
-    constructor(private _navParams: NavParams,
+
+    constructor(
+        private _navParams: NavParams,
         private _listaDeseosService: ListaDeseosService,
-        private _navController: NavController) {
+        private _navController: NavController,
+        public alertCtrl: AlertController
+    ) {
 
     }
 
@@ -32,7 +37,31 @@ export class DetalleComponent implements OnInit {
     }
 
     eliminarLista(): void {
-        this._listaDeseosService.eliminarLista(this.index);
-        this._navController.pop();
+        let confirm = this.alertCtrl.create({
+            title: this.lista.nombre,
+            message: 'Esta seguro que desea borrar la lista?',
+            buttons: [
+                {
+                    text: 'Cancelar'/*,
+                    handler: () => {
+                        console.log('Disagree clicked');
+                    }*/
+                },
+                {
+                    text: 'Eliminar',
+                    handler: () => {
+
+                        this._listaDeseosService.eliminarLista(this.index);
+                        this._navController.pop();
+
+                    }
+                }
+            ]
+        });
+        confirm.present();
     }
+
+
+
+
 }
